+++ include/D2DWnd.h
@@ -40,7 +40,7 @@ namespace YUI
 		virtual void                    OnFinalMessage(HWND hWnd);
 	public:
 		virtual void					Init();
-		virtual void					OnResize(unsigned int width,unsigned int height);
+		virtual void OnResize(float width,float height);
          virtual LRESULT                OnSize(UINT uMsg, WPARAM wParam, LPARAM lParam, BOOL& bHandled);
 		virtual void					OnRender();
 		void							CreateDeviceDependResource();
